+++ products/referrals/backend/temporal/workflow.py
@@ -9,12 +9,13 @@ from datetime import timedelta
 from temporalio import common, workflow
 
 from posthog.temporal.common.base import PostHogWorkflow
-from posthog.temporal.social_referral_status.activities import (
+
+from products.referrals.backend.temporal.activities import (
     referral_status_list_pending_ingestion_activity,
     referral_status_process_single_ingestion_activity,
     referral_status_record_ingestion_check_failure_activity,
 )
-from posthog.temporal.social_referral_status.types import (
+from products.referrals.backend.temporal.types import (
     ProcessSingleReferralIngestionInput,
     RecordIngestionCheckFailureInput,
     SocialReferralStatusInputs,
